// src/api/types/studioType.ts
export interface OrderListRequest {
    /**
     * 列表 or  导出type  0 - 列表  1 - 导出
     */
    type?: string;
    /**
     * 工单状态：1-确认下单信息，2-分配执行人，3-录入执行计划，4-确认大纲，5-确认脚本，6-确认视频初稿，7-发布视频，8-确认收付款，9-财务核账
     */
    orderStatus?: string;
    /**
     * 工商单号
     */
    orderNum?: string;
    /**
     * 订单类型：0-视频工单，1-特殊工单
     */
    orderType?: string;
    /**
     * 签约团队id
     */
    signGroupId?: string;
    /**
     * 签约小组id
     */
    signTeamId?: string;
    /**
     * 签约经纪人id
     */
    signUserId?: string;
    /**
     * 平台id
     */
    platId?: string;
    /**
     * 账号id
     */
    accountId?: string;
    /**
     * 账号名
     */
    accountName?: string;
    /**
     * 品牌id
     */
    brandId?: string;
    /**
     * 收付款状态 1-向达人应付 2-向达人应收
     */
    paymentType?: string;
    /**
     * 收款状态 1-已付款/已收款 2-未付款/未收款
     */
    paymentResult?: string;
    /**
     * 内容协助  0-否 1-是
     */
    contentAssist?: string;
    /**
     * 商务团队id
     */
    crmGroupId?: string;
    /**
     * 商务小组id
     */
    crmTeamId?: string;
    /**
     * 商务id
     */
    crmId?: string;
    /**
     * 档期开始时间
     */
    slotDateBegin?: string;
    /**
     * 档期结束时间
     */
    slotDateEnd?: string;
    /**
     * 绩效月开始时间
     */
    monthMoneyBegin?: string;
    /**
     * 绩效月结束时间
     */
    monthMoneyEnd?: string;
    /**
     * 订单归属flag 1- 好莱坞工单 2- 五合镖局工单
     */
    orderAttributionFlag?: string;
    /**
     * 页码
     */
    pageNum?: string;
    /**
     * 大小
     */
    size?: string;
    /**
     * 个人/团队的标志
     */
    methodId?: string;
    determined?: string;
    "accountIdList[0]"?: string;
    /**
     * 账号标签
     */
    "accountTag[0]"?: string;
    "userIdList[0]"?: string;
    "platList[0]"?: string;
    "orgIdList[0]"?: string;
    /**
     * 平台id
     */
    "platInfos[0].platId"?: string;
    /**
     * 平台名称
     */
    "platInfos[0].platName"?: string;
    /**
     * 组织id
     */
    "orgInfos[0].orgId"?: string;
    /**
     * 组织名
     */
    "orgInfos[0].orgName"?: string;
    /**
     * 组织父id
     */
    "orgInfos[0].parentId"?: string;
    /**
     * 账号id
     */
    "accountInfos[0].accountId"?: string;
    /**
     * 账号名
     */
    "accountInfos[0].accountName"?: string;
    /**
     * 账号统一名称
     */
    "accountInfos[0].belongName"?: string;
    /**
     * 账号id
     */
    "userInfos[0].accountId"?: string;
    /**
     * 账号名
     */
    "userInfos[0].accountName"?: string;
    /**
     * 账号统一名称
     */
    "userInfos[0].belongName"?: string;
    /**
     * 操作方式:1-请求列表2-导出数据
     */
    queryMethodId?: string;
    /**
     * 是否显示已履约订单
     */
     performFlag?: number;
    /**
     * 数据域基础实体
     */
    userIdSQL?: string;
    orgIdSQL?: string;
    accountIdSQL?: string;
    userInfoSQL?: string;
    orgInfoSQL?: string;
    accountInfoSQL?: string;
    platInfoSQL?: string;
  }
  
  /**
   * 接口 [列表接口↗](http://zapi.zhuanspirit.com/project/3631/interface/api/1967490) 的 **返回类型**
   *
   * @分类 [履约流程 - 签约订单模块↗](http://zapi.zhuanspirit.com/project/3631/interface/api/cat_131582)
   * @请求头 `GET /api/sign/order/list`
   * @更新时间 `2022-08-22 17:15:26`
   */
  export interface OrderListResponse {
    /**
     * 合作总额
     */
    allCooperationAmount?: number;
    /**
     * 执行金额
     */
    allExecuteMoney?: number;
    /**
     * 工单实际营收
     */
    orderActualIncomeSum?: number;
     /**
      * 绩效营收
      */
    performanceMoney?: number;
    /**
     * 签约订单列表
     */
    groupOrderInfoBOList: {
      /**
       * 工单号
       */
      orderNo: string;
      /**
       * 品牌名称
       */
       brandName?: string;
      /**
       * 商单类型 1 客户自行下单  2 达人私单  3 不走平台的私单
       */
      quotationType?: number;
      /**
       * 工单类型类型：1=视频工单、2=特殊工单
       */
      orderType?: number;
      /**
       * 工单状态：1-确认下单信息，2-分配执行人，3-录入执行计划，4-确认大纲，5-确认脚本，6-确认视频初稿，7-发布视频，8-确认收付款，9-财务核账
       */
      orderStatus?: number;
      /**
       * 创建人
       */
      createUserName?: number;
      /**
       * 商务团队名
       */
      belongGroupName?: string;
      /**
       * 订单归属
       */
      belongTeam?: string;
      bindUserName?:string;
      /**
       * 账号名
       */
      accountName?: string;
      /**
       * 账号id
       */
      accountId?: number;
      /**
       * 后台登录链接
       */
      accountLoginUrl?: string;
      /**
       * 星图链接
       */
      xingtuIndexUrl?: string;
      /**
       * 标签
       */
      sysAccountTag?: string;
      /**
       * 平台id
       */
      platId?: number;
      /**
       * 平台名
       */
      platName?: string;
      /**
       * 档期
       */
      slotDate?: string;
      /**
       * 视频id
       */
      flowId?: number;
      /**
       * 视频链接
       */
      url?: string;
      /**
       * 视频标题
       */
      titleName?: string;
      /**
       * 视频封面链接
       */
      videoCoverUrl?: string;
      /**
       * 工单实际营收
       */
       orderActualIncome?: number;
       /**
        * 绩效营收
        */
        performanceMoney?: string;
    }[];
    /**
     * 总条数
     */
    total: number;
  }
// ...
  export interface OrderConditionResponse {
    /**
     * 签约筛选联动
     */
    signGroup?: {
      id?: number;
      parentId?: number;
      isParent?: number;
      orgName?: string;
      depthLayer?: number;
      childOrgList?: {
        id?: number;
        parentId?: number;
        isParent?: number;
        orgName?: string;
        depthLayer?: number;
        childOrgList?: {
          id?: number;
          parentId?: number;
          isParent?: number;
          orgName?: string;
          depthLayer?: number;
        }[];
        /**
         * 区分用户还是组织  0 否 1 是
         */
        userFlag?: number;
      }[];
      /**
       * 区分用户还是组织  0 否 1 是
       */
      userFlag?: number;
    }[];
    /**
     * 品牌信息
     */
    brandInfoBOS?: {
      /**
       * id
       */
      id?: number;
      /**
       * 品牌名
       */
      brandName?: string;
      /**
       * 公司名
       */
      companyName?: string;
    }[];
    /**
     * 账号名称
     */
    accountInfo?: {
      accountId?: number;
      accountName?: string;
      belongName?: string;
      platId?: number;
    }[];
    /**
     * 平台信息
     */
    platInfos?: {
      /**
       * 平台id
       */
      platId?: number;
      /**
       * 平台名称
       */
      platName?: string;
    }[];
  }